restructured and refactored project, added integration with social media logins and mongodb, pair programming with George
