+++ .meteor/packages
@@ -20,3 +20,16 @@ insecure@1.0.7                # Allow all DB writes from clients (for prototypin
 static-html             # Define static page content in .html files
 react-meteor-data       # React higher-order component for reactively tracking Meteor data
 bootstrap
+accounts-ui
+accounts-password
+accounts-facebook
+service-configuration
+facebook-config-ui
+accounts-google
+google-config-ui
+accounts-twitter
+accounts-github
+accounts-weibo
+github-config-ui
+twitter-config-ui
+weibo-config-ui
