+++ .meteor/versions
@@ -1,3 +1,13 @@
+accounts-base@1.4.4
+accounts-facebook@1.3.2
+accounts-github@1.4.2
+accounts-google@1.3.3
+accounts-oauth@1.1.16
+accounts-password@1.5.1
+accounts-twitter@1.4.2
+accounts-ui@1.3.1
+accounts-ui-unstyled@1.4.2
+accounts-weibo@1.3.2
 allow-deny@1.1.0
 autopublish@1.0.7
 autoupdate@1.6.0
@@ -5,6 +15,7 @@ babel-compiler@7.3.4
 babel-runtime@1.3.0
 base64@1.0.11
 binary-heap@1.0.11
+blaze@2.3.3
 blaze-tools@1.0.10
 boilerplate-generator@1.6.0
 bootstrap@1.0.1
@@ -15,6 +26,7 @@ check@1.3.1
 ddp@1.4.0
 ddp-client@2.3.3
 ddp-common@1.4.0
+ddp-rate-limiter@1.0.7
 ddp-server@2.3.0
 deps@1.0.12
 diff-sequence@1.1.1
@@ -24,18 +36,28 @@ ecmascript-runtime@0.7.0
 ecmascript-runtime-client@0.8.0
 ecmascript-runtime-server@0.7.1
 ejson@1.1.0
+email@1.2.3
 es5-shim@4.8.0
+facebook-config-ui@1.0.2
+facebook-oauth@1.6.0
 fetch@0.1.1
 geojson-utils@1.0.10
+github-config-ui@1.0.1
+github-oauth@1.2.2
+google-config-ui@1.0.1
+google-oauth@1.2.6
 hot-code-push@1.0.4
 html-tools@1.0.11
 htmljs@1.0.11
+http@1.4.2
 id-map@1.1.0
 insecure@1.0.7
 inter-process-messaging@0.1.0
-jquery@1.0.2
+jquery@1.11.9
 launch-screen@1.1.1
+less@2.8.0
 livedata@1.0.18
+localstorage@1.2.0
 logging@1.1.20
 meteor@1.9.3
 meteor-base@1.4.0
@@ -51,24 +73,44 @@ mongo@1.6.2
 mongo-decimal@0.1.1
 mongo-dev-server@1.1.0
 mongo-id@1.0.7
+npm-bcrypt@0.9.3
 npm-mongo@3.1.2
+oauth@1.2.8
+oauth1@1.2.2
+oauth2@1.2.1
+observe-sequence@1.0.16
 ordered-dict@1.1.0
 promise@0.11.2
 random@1.1.0
+rate-limit@1.0.9
 react-meteor-data@0.2.16
+reactive-dict@1.3.0
 reactive-var@1.0.11
 reload@1.3.0
 retry@1.1.0
 routepolicy@1.1.0
+service-configuration@1.0.11
+session@1.2.0
+sha@1.0.9
 shell-server@0.4.0
 socket-stream-client@0.2.2
+spacebars@1.0.15
 spacebars-compiler@1.1.3
+srp@1.0.12
 standard-minifier-css@1.5.3
 standard-minifier-js@2.4.1
 static-html@1.2.2
+templating@1.2.15
+templating-compiler@1.2.15
+templating-runtime@1.2.15
 templating-tools@1.1.2
 tmeasday:check-npm-versions@0.3.2
 tracker@1.2.0
+twitter-config-ui@1.0.0
+twitter-oauth@1.2.0
 underscore@1.0.10
+url@1.2.0
 webapp@1.7.4
 webapp-hashing@1.0.9
+weibo-config-ui@1.0.1
+weibo-oauth@1.2.1
