--- imports/api/links.js
@@ -1,3 +1,0 @@
-import { Mongo } from 'meteor/mongo';
-
-export default Links = new Mongo.Collection('links');
