+++ imports/api/reservations.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const Reservations = new Mongo.Collection('reservations');
