+++ imports/api/restaurants.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const Restaurants = new Mongo.Collection('restaurants');
