+++ imports/startup/accounts-config.js
@@ -1,0 +1,5 @@
+import { Accounts } from 'meteor/accounts-base';
+
+Accounts.ui.config({
+  passwordSignupFields: 'USERNAME_ONLY'
+});
