--- imports/ui/components/Home.jsx
@@ -1,12 +1,0 @@
-import React, { Component } from 'react'
-import Body from './Body/Body'
-
-export default class Home extends Component {
-  render() {
-    return (
-      <React.Fragment>
-        <Body />
-      </React.Fragment>
-    )
-  }
-}
