+++ imports/ui/components/Pages/Homepage/Slideshow/Slide.jsx
