+++ imports/ui/css/ui.css
@@ -10,24 +10,3 @@
 .form-control {
   margin-top: 5px;
 }
-
-.login-form {
-  width: 100%;
-  max-width: 330px;
-  padding: 15px;
-  margin: auto;
-  height: 100%;
-}
-
-.login-box {
-width: 100%;
-max-width: 330px;
-border-width:2px;
-border-style:solid;
-border-color:black;
-}
-
-.modal {
-  margin: 0;
-  text-align: center;
-}
