+++ server/main.js
@@ -1,5 +1,5 @@
 import { Meteor } from 'meteor/meteor';
-import Links from '/imports/api/links';
+import Restaurants from '/imports/api/restaurants';
 
 function insertLink(title, url) {
   Links.insert({ title, url, createdAt: new Date() });
@@ -7,25 +7,26 @@ function insertLink(title, url) {
 
 Meteor.startup(() => {
   // If the Links collection is empty, add some data.
-  if (Links.find().count() === 0) {
-    insertLink(
-      'Do the Tutorial',
-      'https://www.meteor.com/tutorials/react/creating-an-app'
-    );
 
-    insertLink(
-      'Follow the Guide',
-      'http://guide.meteor.com'
-    );
-
-    insertLink(
-      'Read the Docs',
-      'https://docs.meteor.com'
-    );
-
-    insertLink(
-      'Discussions',
-      'https://forums.meteor.com'
-    );
-  }
+  /*  if (Links.find().count() === 0) {
+      insertLink(
+        'Do the Tutorial',
+        'https://www.meteor.com/tutorials/react/creating-an-app'
+      );
+  
+      insertLink(
+        'Follow the Guide',
+        'http://guide.meteor.com'
+      );
+  
+      insertLink(
+        'Read the Docs',
+        'https://docs.meteor.com'
+      );
+  
+      insertLink(
+        'Discussions',
+        'https://forums.meteor.com'
+      );
+    } */
 });
